+++ lib/src/elements/sink/muxer/mod.rs
@@ -1,9 +1,11 @@
 mod file_muxer;
 mod hls_muxer;
+mod rtmp_muxer;
 mod segmented_file_muxer;
 
 pub use file_muxer::{FileMuxer, FileMuxerError, FileMuxerStreamSink};
 pub use hls_muxer::{
     HlsMode, HlsMuxer, HlsMuxerError, HlsMuxerStreamSink, HlsOptions, HlsSegmentFormat,
 };
+pub use rtmp_muxer::{RtmpMuxer, RtmpMuxerError, RtmpMuxerStreamSink};
 pub use segmented_file_muxer::{SegmentPolicy, SegmentedFileMuxer};